Clean up Variable impl
