+++ SeQuant/core/expressions/variable.hpp
@@ -2,16 +2,15 @@
 #define SEQUANT_EXPRESSIONS_VARIABLE_HPP
 
 #include <SeQuant/core/expressions/expr.hpp>
-#include <SeQuant/core/expressions/expr_ptr.hpp>
 #include <SeQuant/core/expressions/labeled.hpp>
-#include <SeQuant/core/hash.hpp>
 #include <SeQuant/core/utility/macros.hpp>
-#include <SeQuant/core/utility/string.hpp>
 
 #include <string>
 #include <string_view>
 
 namespace sequant {
+
+class ExprPtr;
 
 /// This is represented as a "run-time" complex rational number
 class Variable : public Expr, public MutatableLabeled {
@@ -29,10 +28,9 @@ class Variable : public Expr, public MutatableLabeled {
              std::constructible_from<std::wstring, U>)
   explicit Variable(U &&label) : label_(std::forward<U>(label)) {}
 
-  Variable(std::wstring label) : label_(std::move(label)), conjugated_(false) {}
+  Variable(std::wstring label);
 
-  Variable(const std::string &label)
-      : label_(sequant::toUtf16(label)), conjugated_(false) {}
+  Variable(const std::string &label);
 
   /// @return variable label
   /// @warning conjugation does not change it
@@ -48,9 +46,9 @@ class Variable : public Expr, public MutatableLabeled {
 
   std::wstring to_latex() const override;
 
-  type_id_type type_id() const override { return get_type_id<Variable>(); }
+  type_id_type type_id() const override;
 
-  bool is_scalar() const override { return true; }
+  bool is_scalar() const override;
 
   ExprPtr clone() const override;
 
@@ -61,26 +59,10 @@ class Variable : public Expr, public MutatableLabeled {
   std::wstring label_;
   bool conjugated_ = false;
 
-  hash_type memoizing_hash() const override {
-    auto compute_hash = [this]() {
-      auto val = hash::value(label_);
-      hash::combine(val, conjugated_);
-      return val;
-    };
+  hash_type memoizing_hash() const override;
 
-    if (!hash_value_) {
-      hash_value_ = compute_hash();
-    } else {
-      SEQUANT_ASSERT(*hash_value_ == compute_hash());
-    }
+  bool static_equal(const Expr &that) const override;
 
-    return *hash_value_;
-  }
-
-  bool static_equal(const Expr &that) const override {
-    return label_ == static_cast<const Variable &>(that).label_ &&
-           conjugated_ == static_cast<const Variable &>(that).conjugated_;
-  }
 };  // class Variable
 
 }  // namespace sequant
